update init_path in behaviour
