+++ ml-agents/mlagents/trainers/settings.py
@@ -752,7 +752,6 @@ class CheckpointSettings:
     train_model: bool = parser.get_default("train_model")
     inference: bool = parser.get_default("inference")
     results_dir: str = parser.get_default("results_dir")
-    init_checkpoints_list: Optional[Dict[str, str]] = None
 
     @property
     def write_path(self) -> str:
